cleanup:
- use central package management
- removed unused code
- moved files in proper namespaces
- fix some warnings
