+++ Directory.Build.props
@@ -1,3 +1,8 @@
 <Project>
-  <Import Project="build\Common.props" />
+  <PropertyGroup>
+    <Authors>Quotation Factory</Authors>
+    <Product>Integration Quickstart</Product>
+    <RepositoryUrl>https://github.com/QuotationFactory/Integration.Quickstart</RepositoryUrl>
+    <RepositoryType>git</RepositoryType>
+  </PropertyGroup>
 </Project>
