--- Directory.Build.targets
@@ -1,3 +1,0 @@
-<Project>
-  <Import Project="build\Common.targets" />
-</Project>
