+++ Integration.Common/FileWatcher/BufferingFileSystemWatcher.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Concurrent;
+﻿#nullable disable
+using System.Collections.Concurrent;
 using System.Collections.ObjectModel;
 using System.ComponentModel;
