--- Integration.Common/FileWatcher/FileWatcherErrorEventArgs.cs
@@ -1,12 +1,0 @@
-﻿using System.ComponentModel;
-
-namespace Integration.Common.FileWatcher;
-
-public class FileWatcherErrorEventArgs : HandledEventArgs
-{
-    public readonly Exception Error;
-    public FileWatcherErrorEventArgs(Exception exception)
-    {
-        this.Error = exception;
-    }
-}
