--- Integration.Common/FileWatcher/IFileWatcherSetting.cs
@@ -1,7 +1,0 @@
-﻿namespace Integration.Common.FileWatcher;
-
-public interface IFileWatcherSetting
-{
-    string Directory { get; set; }
-    IEnumerable<string> Filters { get; set; }
-}
