--- Integration.Common/Notifications/IntegrationOutputFileCreated.cs
@@ -1,13 +1,0 @@
-﻿using MediatR;
-
-namespace Integration.Common.Notifications;
-
-public class IntegrationOutputFileCreated : INotification
-{
-    public string FilePath { get; }
-
-    public IntegrationOutputFileCreated(string filePath)
-    {
-        FilePath = filePath;
-    }
-}
