+++ Integration.Host/Features/OutputFile/OutputFileCreatedHandler.cs
@@ -5,8 +5,9 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using ICSharpCode.SharpZipLib.Zip;
-using Integration.Common.Classes;
-using Integration.Common.Extensions;
+using Integration.Common;
+using Integration.Common.Serialization;
+using Integration.Host.Configuration;
 using MediatR;
 using MetalHeaven.Agent.Shared.External.Interfaces;
 using MetalHeaven.Agent.Shared.External.Messages;
