+++ Integration.Host/Properties/launchSettings.json
@@ -1,6 +1,6 @@
 ﻿{
   "profiles": {
-    "Metal.Heaven.Integration.Host": {
+    "Integration.Host": {
       "commandName": "Project",
       "dotnetRunMessages": "true",
       "environmentVariables": {
